+++ src/common/ui/index.tsx
@@ -1,0 +1,11 @@
+import { styled } from "@stitches/react";
+
+export const Icon = styled("img", {
+  variants: {
+    size: {
+      big: {
+        width: "45px"
+      }
+    }
+  },
+});
