+++ src/page/home.tsx
@@ -1,5 +1,5 @@
-import Header from "../components/header/header";
-import Nav from "../components/nav/nav";
+import Header from "components/header";
+import Nav from "components/nav";
 
 const Home = () => {
   return (
